backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
httpx==0.24.1
python-multipart==0.0.9
pydantic-settings==2.1.0
python-dotenv==1.0.1
gunicorn==21.2.0

supabase==2.3.0
pyjwt==2.8.0
